+++ simple_history/models.py
@@ -41,9 +41,10 @@ class HistoricalRecords(object):
     thread = threading.local()
 
     def __init__(self, verbose_name=None, bases=(models.Model,),
-                 user_related_name='+'):
+                 user_related_name='+', table_name=None):
         self.user_set_verbose_name = verbose_name
         self.user_related_name = user_related_name
+        self.table_name = table_name
         try:
             if isinstance(bases, six.string_types):
                 raise TypeError
@@ -114,6 +115,8 @@ class HistoricalRecords(object):
         attrs.update(self.get_extra_fields(model, fields))
         # type in python2 wants str as a first argument
         attrs.update(Meta=type(str('Meta'), (), self.get_meta_options(model)))
+        if self.table_name is not None:
+            attrs['Meta'].db_table = self.table_name
         name = 'Historical%s' % model._meta.object_name
         registered_models[model._meta.db_table] = model
         return python_2_unicode_compatible(