+++ simple_history/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-__version__ = '1.6.3'
+__version__ = '1.7.0'
 
 
 def register(