+++ CHANGES.rst
@@ -5,6 +5,7 @@ tip (unreleased)
 ----------------
 - Add ability to list history in admin when the object instance is deleted. (gh-72)
 - Add ability to change history through the admin. (Enabled with the `SIMPLE_HISTORY_EDIT` setting.)
+- Add Django 1.9 support.
 
 1.6.3 (2015-07-30)
 ------------------