+++ CHANGES.rst
@@ -6,6 +6,7 @@ tip (unreleased)
 - Add ability to list history in admin when the object instance is deleted. (gh-72)
 - Add ability to change history through the admin. (Enabled with the `SIMPLE_HISTORY_EDIT` setting.)
 - Add Django 1.9 support.
+- Support for custom tables names. (gh-196)
 
 1.6.3 (2015-07-30)
 ------------------